+++ smart-contracts/wasm-chain-integration/src/v0/ffi.rs
@@ -1,5 +1,10 @@
 use crate::v0::*;
-use concordium_wasm::{artifact::CompiledFunctionBytes, output::Output, utils::parse_artifact};
+use concordium_wasm::{
+    artifact::CompiledFunctionBytes,
+    output::Output,
+    utils::{parse_artifact, InstantiatedModule},
+    validate::ValidationConfig,
+};
 use libc::size_t;
 
 /// All functions in this module operate on serialized artifact bytes. For
@@ -188,10 +193,14 @@ unsafe extern "C" fn validate_and_process_v0(
 ) -> *mut u8 {
     let wasm_bytes = slice_from_c_bytes!(wasm_bytes_ptr, wasm_bytes_len);
     match utils::instantiate_with_metering::<ProcessedImports, _>(
+        ValidationConfig::V0,
         &ConcordiumAllowedImports,
         wasm_bytes,
     ) {
-        Ok(artifact) => {
+        Ok(InstantiatedModule {
+            custom_sections_size: _,
+            artifact,
+        }) => {
             let mut out_buf = Vec::new();
             let num_exports = artifact.export.len(); // this can be at most MAX_NUM_EXPORTS
             out_buf.extend_from_slice(&(num_exports as u16).to_be_bytes());